+++ app/Http/Livewire/IdeaShow.php
@@ -4,6 +4,7 @@ namespace App\Http\Livewire;
 
 use App\Exceptions\DuplicateVoteException;
 use App\Exceptions\VoteNotFoundException;
+use App\Models\Comment;
 use App\Models\Idea;
 use Illuminate\Support\Facades\DB;
 use Livewire\Component;
@@ -12,13 +13,23 @@ class IdeaShow extends Component
 {
     public $idea;
     public $votesCount;
+    public $commentsCount;
     public $hasVoted;
+    public $commentBody;
 
-    public function mount(Idea $idea, $votesCount)
+    protected $listeners = ['comment-deleted' => 'decrementTheCommentCount'];
+
+    public function mount(Idea $idea, $votesCount,$commentsCount)
     {
         $this->idea = $idea;
         $this->votesCount = $votesCount;
+        $this->commentsCount = $commentsCount;
         $this->hasVoted = $idea->isVotedByUser(auth()->user());
+    }
+
+    public function decrementTheCommentCount()
+    {
+        $this->commentsCount--;
     }
 
     public function vote()
@@ -51,8 +62,39 @@ class IdeaShow extends Component
         return view('livewire.idea-show');
     }
 
+    public function postComment()
+    {
+        if (! auth()->check()) {
+            return redirect(route('login'));
+        }
+
+        $this->validate([
+            'commentBody' => 'required',
+        ]);
+
+        Comment::create([
+            'user_id' => auth()->user()->id,
+            'idea_id' => $this->idea->id,
+            'body' => $this->commentBody,
+        ]);
+        
+        $this->commentBody = '';
+        $this->commentsCount++;
+        $this->emit('comment-saved');
+    }
+
     public function deleteIdea(Idea $idea)
     {
+        if (! auth()->check()) {
+            return redirect(route('login'));
+        }
+
+        if ($idea->comments->count() > 3) {
+            session()->flash('error', 'You cannot delete this idea because it got more then 3 Comments.');
+
+            return back();
+        }
+
         DB::table('votes')->where('idea_id',$idea->id)->delete();
 
         if($idea->delete())
